Add descriptive prints, more specific error handling, fix syntax.
Added descriptive prints to  inform the user of the progress of the
script.

Added more specific error handling instead of the general “Exception”.
Also added DryRun option to the add/remove modify_image_attribute and
the logic to handle it.

Fixed json_data function’s return. It wasn’t in the correct space.